# query by pool instead of using the connection directly