@ericdum 2021-05-17
This PR is intended to increase the performance of **Concurrent Request** and to resolve some errors like **CONNREST**. 
To do that by let MySQL Pool manage the connections instead of manually manage connections.