@dceejay 2021-05-17
Hi,

thanks for this - a couple of points.. 1) are you able to accept/click/sign the CLA so we can merge it ?
and 2) does this handle connection errors OK - There has been along standing issue of connections seeming to timeout and/or drop over a long time - and not reconnect... Does this handle this somehow ? I can' see any error handler for he pool ?