@ericdum 2021-05-18
Hi @dceejay 
1) I have been signed the CLA, but the checking program seems not to recheck. I don't know how to make it work.
2) After switching the management to the pool, most of the errors will be handled by the pool also. So, there's only error left to the program is **pool.query(err, ....)**, so I have already handled it.

And FYI, there's some benchmark result below.
**Before optimize:** ab -n 500 -c 30
```
Document Path:          /data-flow/api/benchmark
Document Length:        13 bytes

Concurrency Level:      30
Time taken for tests:   13.528 seconds
Complete requests:      500
Failed requests:        0
Total transferred:      138500 bytes
HTML transferred:       6500 bytes
Requests per second:    36.96 [#/sec] (mean)
Time per request:       811.655 [ms] (mean)
Time per request:       27.055 [ms] (mean, across all concurrent requests)
Transfer rate:          10.00 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:       55  693 2318.4     88    9888
Processing:    59  102  34.5     93     204
Waiting:       59  102  34.5     93     203
Total:        117  795 2316.3    189    9995

Percentage of the requests served within a certain time (ms)
  50%    189
  66%    210
  75%    239
  80%    271
  90%    500
  95%   9902
  98%   9994
  99%   9994
 100%   9995 (longest request)```

**After optimize:** ab -n 2000 -c 30
```
Document Path:          /data-flow/api/benchmark
Document Length:        13 bytes

Concurrency Level:      30
Time taken for tests:   2.299 seconds
Complete requests:      2000
Failed requests:        0
Total transferred:      554000 bytes
HTML transferred:       26000 bytes
Requests per second:    869.93 [#/sec] (mean)
Time per request:       34.485 [ms] (mean)
Time per request:       1.150 [ms] (mean, across all concurrent requests)
Transfer rate:          235.32 [Kbytes/sec] received

Connection Times (ms)
              min  mean[+/-sd] median   max
Connect:        7   16   4.4     15      39
Processing:     8   18   4.7     17      53
Waiting:        8   18   4.7     17      53
Total:         19   34   7.2     32      64

Percentage of the requests served within a certain time (ms)
  50%     32
  66%     35
  75%     37
  80%     39
  90%     43
  95%     48
  98%     52
  99%     57
 100%     64 (longest request)
```